fix(ui): fix date tile
