+++ components/layout.module.scss
@@ -17,10 +17,12 @@
 }
 
 .bgColor {
+  width: 500%;
+  transform: translate3d(-35%, -10%, 0px);
+  z-index: 0;
   background-color: $active-color;
   background-size: 45px 45px;
-  background-image:
-    linear-gradient(to right, #7a00f3 1px, transparent 1px),
+  background-image: linear-gradient(to right, #7a00f3 1px, transparent 1px),
     linear-gradient(to bottom, #7a00f3 1px, transparent 1px);
   border: 1.1rem solid white;
   padding: 1.5rem;
